+++ src/module/applications/apps/ability-configuration-dialog.mjs
@@ -19,6 +19,7 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
     actions: {
       panToken: this.#panToken,
       placeTemplate: this.#placeTemplate,
+      swapMeleeRanged: this.#swapMeleeRanged,
     },
   };
 
@@ -179,6 +180,11 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
 
     context.resource.show = this.item.system.resource;
 
+    context.meleeRanged = {
+      show: this.item.system.distance.type === "meleeRanged",
+      isMelee: this.item.system.damageDisplay === "melee",
+    };
+
     // Heroic resource/malice spend
     if (this.item.system.effects.documentsByType.spend.length) {
       context.resource.show = true;
@@ -289,5 +295,18 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
   static async #placeTemplate(event, target) {
     if (this.#region) await this.#region.delete();
     this.#region = await this.item.system.placeTemplate({ setTargets: "replace" });
+  }
+
+  /* -------------------------------------------------- */
+
+  /**
+   * Switch if this ability is going to roll as melee or ranged.
+   * @this AbilityConfigurationDialog
+   * @param {PointerEvent} event   The originating click event.
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action].
+   */
+  static async #swapMeleeRanged(event, target) {
+    await this.item.update({ "system.damageDisplay": this.item.system.damageDisplay === "melee" ? "ranged" : "melee" });
+    this.render({ parts: ["ability"] });
   }
 }